requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
inflection==0.5.1
Markdown==3.6
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
